[Step4] fix : step3 리뷰 반영 - 승, 패 판정에 대한 클래스 생성
