+++ src/main/kotlin/blackjack/domain/game/BlackJackGame.kt
@@ -27,43 +27,4 @@ class BlackJackGame(
             it.addCard(cardDeck.pickCard())
         }
     }
-
-    fun calculateWinner() {
-        _players.players
-            .forEach {
-                judge(it, _dealer)
-            }
-    }
-
-    private fun judge(player: Player, dealer: Dealer) {
-        if (checkPlayerWin(player, dealer)) {
-            player.isWinner = true
-            dealer.lose++
-        }
-
-        if (checkDealerWin(player, dealer)) {
-            player.isWinner = false
-            dealer.win++
-        }
-    }
-
-    private fun checkPlayerWin(player: Player, dealer: Dealer): Boolean {
-        if (player.score > BLACKJACK_SCORE) {
-            return false
-        }
-
-        return dealer.score > BLACKJACK_SCORE || player.score > dealer.score || player.score == BLACKJACK_SCORE
-    }
-
-    private fun checkDealerWin(player: Player, dealer: Dealer): Boolean {
-        if (dealer.score > BLACKJACK_SCORE) {
-            return false
-        }
-
-        return player.score > BLACKJACK_SCORE || player.score < dealer.score || dealer.score == BLACKJACK_SCORE
-    }
-
-    companion object {
-        private const val BLACKJACK_SCORE = 21
-    }
 }
